createPao {RPresence}	R Documentation

Create a pao data object.

Description

createPao Creates a pao data object from R variables.

Usage

createPao(data, unitcov = NULL, survcov = NULL,
  nsurveyseason = ncol(data), nmethods = 1, frq = rep(1, nrow(data)),
  title = "PRESENCE Analysis", unitnames = paste0("unit", 1:nrow(data)),
  paoname = "pres.pao")

Arguments

data
data frame containing the detection (=1) and nondetection (=0) data with observations from a sampling unit along a row, and surveys in the columns. Missing values are allowed (=NA).

unitcov
data frame for unit-specific covariates (e.g., habitat, elevation, etc.) with 1 row per unit (so nrow(unitcov) == nrow(data)). Each column of the data frame should contain a single covariate, with the column name used as the covariate name. Covariatecolumn names shoud not contain spaces, and avoid using all uppercase as some terms are used by the package to define common effects used in models, e.g., SURVEY. If there are no unit-specific covariates then unitcov=NULL. Note that additional unit-specific covariates can be supplied when fitting a model with occMod.

survcov
data frame for survey-specific covariates (e.g., survey time, weather conditions, observer, etc.) with 1 row per survey (so nrow(survcov) == nrow(data)*ncol(data)). Each column of the data frame should contain a single covariate, with the column name used as the covariate name. Covariate/column names shoud not contain spaces, and avoid using all uppercase as some terms are used by the package to define common effects used in models, e.g., SURVEY. If there are no survey-specific covariates then survcov=NULL. Note that additional survey-specific covariates can be supplied when fitting a model with occMod. Note that the format of survey-specific covariates in survcov is different from that used in PRESENCE (i.e., long vs flat format respectively.

nsurveyseason
vector of the number of surveys per season (e.g., if three seasons, with 4 surveys per season, nsurveyseason=c(4,4,4)). The default assumes that data is for a single-season or static occupancy model.

nmethods
integer number of methods per survey

title
(optional) a label included in the PRESENCE output files, of little practical consequence.

unitnames
a vector of string containing labels for each sampling unit. Included in the results as row names for the real parameter estimates of the occMod object.

paoname
string containing a name for the temporary pao

Value

list of class "pao" containing the data required to fit an occupancy model using occMod. The list contains the following objects:

nunits
number of units or sampling units, i.e., nrow(data).

nsurveys
total number of surveys, i.e., ncol(data).

nseasons
number of seasons of data.

det.data
detection/nondetection data, i.e., data.

unitcov
see above. If no covariates are supplied a vector of 1's named TEMP is added.

survcov
see above. A factor called SURVEY is automatically added to allow survey-specific detection probabilities (but equal across units).

nsurveyseason
see above.

title
see above.

unitnames
see above.

surveynames
a vector of strings providing labels for the survey occasions. Of little practical consequence.

paoname
see above. Equals f for createPao.

[Package RPresence version 2.12.17 Index]