fitted.occMod {RPresence}	R Documentation

Extract fitted values from occupancy model, i.e., real parameter estimates

Description

Extract fitted values from occupancy model, i.e., real parameter estimates

Usage

## S3 method for class 'occMod'
fitted(object, param = NULL)

Arguments

object
an object of class occMod which is the result from fitting a model using the occMod function.

param
name of real parameter to extract

Value

a data.frame of the estimates, standard errors and confidence intervals limits

Author(s)

Darryl MacKenzie

[Package RPresence version 2.12.17 Index]