genpresEV {RPresence}	R Documentation

generate expected value detection histories

Description

genpresEV generates expected value detection history frequencies for single or multi-season models (false-positives optional)

Usage

genpresEV(N = 100, K = 4, psi = 0.75, p = 0.4, fp = 0, b = 0,
  eps = 0, gam = 0, sps = 4)

Arguments

N
number of sites

K
number of surveys

psi
probability of occupancy

p
probability of detection (not false positive)

fp
probability of false detection

b
probability of "sure" detection

eps
probability of local extinction between seasons

gam
probability of local colonization between seasons

sps
number of surveys per season

Value

list(hst=det. history matrix, frq=vector of frequencies)

Note

This function generates at least 2^K detection-histories, so if K>8, it may take a very long time to run.

Examples

   # generate expected value data for standard single-season model (no false positives)
hstfrq=genpresEV(N=1000, K=4, psi=.8, p=.3)  # generate det. hists and freqs
EVpao=createPao(hstfrq$hst,frq=hstfrq$frq)                  #  create pao file

   # generate expected value data for single-season model with false positive detections
hstfrq=genpresEV(N=1000, K=4, psi=.8, p=.3, fp=0.1, b=0.5)  # generate det. hists and freqs
EVpao=createPao(hstfrq$hst,frq=hstfrq$frq)                  #  create pao file

   # generate expected value data for multi-season model (no false positive detections)
hstfrq=genpresEV(N=1000, K=4, psi=.8, p=.3, eps=.2, gam=.1, sps=2)  # generate det. hists and freqs
EVpao=createPao(hstfrq$hst,frq=hstfrq$frq,nsurveyseason=2)   #  create pao file

[Package RPresence version 2.12.17 Index]