genpresEVms {RPresence}	R Documentation

generate expected value detection histories for multi-state models

Description

genpresEVms generates expected value detection history frequencies for single or multi-season multi-state models.

Usage

genpresEVms(N = 100, K = 4, psi = 0.5, R = 0.8, p1 = 0.4, p2 = 0.3,
  dlta = 0.6, nstates = 3, psi1 = 0, Cpsi = 0, CR = 0, phi = 0,
  p = 0, sps = 4)

Arguments

N
number of sites

K
number of surveys

psi
probability of occupancy

R
probability of breeding (state=2), given occupancy

p1
probability of detection, given occupied (state=1)

p2
probability of detection, given occupied and breeding (state=2)

dlta
probability of detecting breeding, given occupancy state=2

nstates
number of states (including "not occupied state")

psi1
probability vector of initial states (not occ, occ, occ_breeding)

Cpsi
conditional probability of occupancy, given previous state

CR
conditional probability of breeding, given occupancy and previous state

phi
transition probability matrix (an alternative to specifying psi,R,Cpsi,CR)

p
detection probability matrix (an alternative to specifying p1,p2,dlta)

sps
number of surveys per season

Value

list(hst=det. history matrix, frq=vector of frequencies)

Note

This function generates at least nstates^(K-1) detection-histories, so if K>6, it may take a very long time to run.

Examples

   # generate expected value data for standard single-season model (psi,R parameterization)
x1=genpresEVms(N=100,K=4,psi=.75,R=.8,p1=.4,p2=.3,dlta=.6,nstates=3,psi1=0,Cpsi=0,CR=0,phi=0,p=0,sps=4)

   # generate expected value data for multi-season model (psi,R parameterization)
x2=genpresEVms(N=100,K=4,psi=.75,R=.8,p1=.4,p2=.3,dlta=.6,nstates=3,psi1=0,
               Cpsi=c(.66,.1,.15),CR=c(.72,.77,0),phi=0,p=0,sps=2)

   # generate expected value data for multi-season model (phi parameterization)
nstates=3; p=matrix(0,nstates,nstates); p[2,2]=.4; p[3,2]=.3*.4; p[3,3]=.3*.6
psi1=c(.25,.75*(1-.8),.75*.8)
phi=rbind(c(.66*(1-.72),.66*.72),c(.1*(1-.77),.1*.77),c(.15,.0)); phi=cbind(1-rowSums(phi),phi)
x3=genpresEVms(N=100,K=4,nstates=3,psi1=psi1,phi=phi,p=p,sps=2)

   # generate expected value data for multi-season model (phi parameterization)
nstates=3; p=matrix(0,nstates,nstates); p[2,2]=p[3,2]=p[3,3]=.25
psi1=c(.167,.5,.333)
phi=rbind(c(.5,.333),c(.5,.333),c(.5,.333)); phi=cbind(1-rowSums(phi),phi)
x4=genpresEVms(N=100,K=4,nstates=3,psi1=psi1,phi=phi,p=p,sps=2)

   # generate expected value data for multi-season model (phi parameterization)
nstates=4; p=matrix(0,nstates,nstates); p[2:4,2]=p[3:4,3]=p[4,4]=.2
psi1=c(0,.4,.3,.2); psi1[1]=1-sum(psi1)
phi=matrix(0,nstates,nstates); phi[,2]=.4; phi[,3]=.3; phi[,4]=.2; phi[,1]=1-rowSums(phi)
x4=genpresEVms(N=100,K=4,nstates=4,psi1=psi1,phi=phi,p=p,sps=2)
EVpao=createPao(x4$hst,frq=x4$frq,nsurveyseason=2,nstates=4)   #  create pao file

[Package RPresence version 2.12.17 Index]