importProject {RPresence}	R Documentation

Import a PRESENCE project into R.

Description

Imports the results stored in a PRESENCE project folder into R as an aic.table object. Note that in this package version the code has only been developed to import static occupancy/single-season models, and the first and fourth parameterisation of the dynamic occupancy/multi-season models. Because design matrices in PRESENCE are constructed manually, and can be done in multiple ways, no attempt has been made to second-guess the intended model stucture with respect to covariates. Therefore predictions can not be made from imported models (unless manually reconfigured), although model averaged estimates of real parameters of the surveyed units can be obtained.

Usage

importProject(pa3)

Arguments

pa3
a string containing the path (if outside of the current working directory) and file name of the PRESENCE project summary file with the extension .pa3. e.g., import.project("bird_project/bird_project.pa3")

Value

Returns an object of class aic.table, the same as the result of the createAicTable function. The only difference being that the model component of the stored occMod objects = NULL. i.e., the model structure is not imported and left undefined.

Author(s)

Darryl MacKenzie.

See Also

createAicTable,occMod, predict.occMod,modAvg

[Package RPresence version 2.12.17 Index]