modAvg {RPresence}	R Documentation

Calculate model averaged estimates

Description

Calculates model averaged estimates of real parameters stored in a aicTable object; the result of createAicTable. Does not perform model averaging of the regression coefficients, or beta parameters. Can be used to model average predicted probabilities too.

Usage

modAvg(aic.tab, param = NULL, index = 1:nrow(aic.tab$table), conf = 0.95,
  predict = FALSE, newdata)

Arguments

aic.tab
an aicTable object; the result of createAicTable.

param
a string with the name of the real parameter to model average, e.g., param = "psi". Presently, only works for parameters "psi" and "p".

index
a vector of integer values indexing which ranked models are to be model averaged across. Can be used if only model averaging across a subset of models. Note that AIC weights are recalcuated such that their sum = 1 for the subset.

conf
level for confidence interval (may be vector valued).

predict
if FALSE (default) then model averages observed real estimates. Otherwise model averages predicted values for param from the supplied newdata.

newdata
a data frame containing the values of the covariates for which predictions are required. Variable names must exactly match those used in the data. May contain varaibles that are not used in some models.

Value

returns a data frame containing the model averaged estimates, standard errors and limts of a 95% confidence interval.

Author(s)

Darryl MacKenzie

See Also

createAicTable, predict.occMod

[Package RPresence version 2.12.17 Index]