modCombos {RPresence}	R Documentation

Create a list of formulae for all possible first-order combinations for a set of variables.

Description

Creates a list of formulae for the parameter param which contains all possible first-order (i.e., no interactions) combinations for the set of variables cov.

Usage

modCombos(param, covs)

Arguments

param
a string with the name of the parameter of interest, e.g., param="psi".

covs
a vector of string values containing the names of the variables of interest.

Value

returns a list of formulae of each combination of variables.

Author(s)

Darryl MacKenzie.

See Also

createAicTable, predict.occMod

[Package RPresence version 2.12.17 Index]