occMod {RPresence}	R Documentation

Fit an occupancy model using Program PRESENCE

Description

occMod returns a list variable containing detection-history and covariate data.

Usage

occMod(model, cov.list = NULL, data, type = NULL, conf = 0.95,
  modname = NULL, paoname = NULL, outfile = NULL, fixed = NULL,
  initvals = NULL, modfitboot = NULL, VCoutopt = "realbetavc",
  noDerived = F, randinit = 0, maxfn = "32000", ...)

Arguments

model
list of formula for each paramter (eg., psi, p, gam, eps...)

cov.list
list of covariates used in the model

data
pao data object

type
a string indicating the type of occupancy model to fit. Currently implemented type options are:

type 	 Descriptionxxx 	 Function (see for details) 	 Reserved covariate names

"so" 	 Static occupancy, single season 	 occMod_SO	 SURVEY

"so.fp" 	 Static occupancy, single season, false-positive detections	 occMod_SO_fp	 SURVEY

"so.mm" 	 Static occupancy, single season, multi method (or multi-scale) 	 occMod_SO_mm	 SURVEY

"so.cd" 	 Static occupancy, single season, correlated detections 	 occMod_SO_cd	 SURVEY

"2sp"	 Static or Dynamic occupancy, 2 species, "psiBA" or "nu" param 	 occMod_2SP	 SURVEY,SP,INT,INT_o,INT_d,SEASON

"so.2sp.fp"	 Static occupancy, single season, 2 species, false positive detections 	 occMod_SO_2SP_cd	 SURVEY,SP,INT,INT_o,INT_d

"do.1" 	 Dynamic occupancy, multi season, 1st param 	 occMod_DO	 SURVEY,SEASON

"do.fp" 	 Dynamic occupancy, multi season, with false-positive detections 	 occMod_DO_fp	 SURVEY,SEASON

"do.4" 	 Dynamic occupancy, multi season, 4th param 	 occMod_DO4	 SURVEY,SEASON

"do.ms.1"	 Dynamic occupancy, multi season, multi state, 1st param	 occMod_DO_ms1	 SURVEY,SEASON,DYN,PREV_STATE,STATE

"do.ms.2"	 Dynamic occupancy, multi season, multi state, 2nd param	 occMod_DO_ms2	 SURVEY,SEASON,DYN,PREV_STATE,STATE

"do.2sp.1"	 Dynamic occupancy, multi season, 2 species, psiBA param 	 occMod_2SP	 SURVEY,SP,INT,INT_o,INT_d

conf
level for confidence interval (may be vector valued).

modname
(optional) a string with the user supplied name for a model. If NULL, modname is created from the formulae supplied in model.

paoname
(optional) a string with the user supplied filename for PRESENCE data and output files. If NULL, a generated name is used.

fixed
a data.frame with variables param and value. param should be a vector of string values containing the names of the real parameters (eg., "p(2)") to fix, and value be a numeic vector of the values the real parameter should be fixed at. Ordering depends on model type.

initvals
a vector providing initial values for regression coefficients for the optimization procedure used by PRESENCE. Ordering depends on model type.

modfitboot
number of bootstraps for assessing model fit. Only works for type = "so".

VCoutopt
option for PRESENCE computation/output of beta/real var-cov matrices

randinit
number of random initial starting value vectors to try (defalut=0)

...
additional arguments to occMod.XX function.

Details

choices for VCoutopt are:

"nose"only compute likelihood and beta estimates, no beta var-cov matrix or beta SE's (useful for faster model selection of big models or simulations)

"betavc"compute but don't print beta var-cov matrix so we get beta SE's, don't compute real params

"noreal"compute and print beta beta var-cov matrix, don't compute real params (I think this is what you want for RPresence)

"novcs"compute real params and real var-cov matrix, don't print either var-cov matrix.

"realbetavc"(default) compute real params and real var-cov matrix, print only beta var-cov matrix.

"bothvc"print both var-cov matrices.

Value

Returns a list of class "occMod" and class type. The occMod object has the following objects:

modname
the name of the model

model
a named list containing the right-hand formula for each real parameter type.

data
a copy of the temporary pao object that is created from the model argument then used to create the temporary PRESENCE data file for fitting the model.

outfile
name of PRESENCE output file.

neg2loglike
-2*log(likelihood) value for the model evaluated at the maximum likelihood estimates.

npar
number of parameters in the model, calculated as the number of regression coefficients.

aic
value of Akaike's Information Criterion.

beta
list of estimated regression coefficents (sometimes called beta parameters), standard errors and variance-covariance matrix for each parameter type. Exact content depends on the type of model being fit to the data. Values will be on either the logit or log scale depending on whether the parameter type is a probability or not, respectively.

real
list of real parameter estimates on their natural scale (i.e., 0-1 for probabilities). A data frame for each parameter type the contains the estimate, standard error and limits of the requested confidence intervals. Exact content depends on the type of model being fit to the data.

derived
list of parameter estimates that are derived from real parameters and not directly estimated. A data frame for each parameter type the contains the estimate, standard error and limits of the requested confidence intervals. Exact content depends on the type of model being fit to the data.

modfit
list of statistics from parametric bootstrap-based model fit.

warnings
a list containing any warnings from the PRESENCE output file; conv=number of significant digits convergence was achieved to (=NULL if >7); and VC=1 if there is a problem with the variance-covariance matrix (likely non-invertiable observed Hessian matrix), =NULL otherwise. If conv<3 then may want to attempt alternative initial values or rescaling covariates to see if different results are obtained. If VC=1 then all standard errors and confidence intervals should be ignored.

version
a list containing the version numbers for PRESENCE and RPresence used when fitting the model

Author(s)

Darryl MacKenzie

See Also

occMod_SO,occMod_SO_mm,occMod_2SP, occMod_DO,occMod_DO4,occMod_DO_ms2,occMod_2SPfp

Examples

# load a PRESENCE data file
filename<-system.file("extdata/weta.pao",package="RPresence")
weta.data<-readPao(filename)

## convert indicator variables to categorical covariates
weta.data$unitcov$Habitat[weta.data$unitcov$Browsed==1] <- "browsed"
weta.data$unitcov$Habitat[weta.data$unitcov$Unbrowsed==1] <- "unbrowsed"
weta.data$survcov$Obs<-as.factor(
     1*weta.data$survcov$Obs1+
     2*weta.data$survcov$Obs2+
     3*weta.data$survcov$Obs3)

## fit some models
mod1<-occMod(model=list(psi~Habitat,p~SURVEY+Obs),data=weta.data,type="so")
mod2<-occMod(model=list(psi~Habitat,p~Obs),data=weta.data,type="so")
mod3<-occMod(model=list(psi~Habitat,p~SURVEY),data=weta.data,type="so")
mod4<-occMod(model=list(psi~Habitat,p~1),data=weta.data,type="so")
mod5<-occMod(model=list(psi~1,p~SURVEY+Obs),data=weta.data,type="so")
mod6<-occMod(model=list(psi~1,p~Obs),data=weta.data,type="so")
mod7<-occMod(model=list(psi~1,p~SURVEY),data=weta.data,type="so")
mod8<-occMod(model=list(psi~1,p~1),data=weta.data,type="so")
#
## create AIC table
models<-list(mod1,mod2,mod3,mod4,mod5,mod6,mod7,mod8)
results<-createAicTable(models)
summary(results)

## perform model averageing
psi.ma<-modAvg(aic.tab=results,param="psi")
p.ma<-modAvg(aic.tab=results,param="p")

## Not run: 
m1<-occMod(model=list(psi~habitat,gamma~habitat+SEASON,epsilon~habitat+SEASON,
                       p~habitat+SEASON),data=data,type="do.1")

## End(Not run)

[Package RPresence version 2.12.17 Index]