occMod_2SP {RPresence}	R Documentation

Fit two species, static occupancy (single season), or dynamic occupancy (multi-season) model

Description

This is not intended for direct use, but instead the occMod function should be used with type="so.2sp.1" or type="so.2sp.2" depending on the parameterisation required. NOTE THAT THERE MAY BE SOME CHANGES TO HOW THIS MODEL IS IMPLEMENTED IN THE NEAR FUTURE.

Usage

occMod_2SP(psi = call(), psi.cov = data$unitcov, p = call(),
  p.cov = cbind(data$unitcov, data$survcov), gamma = call(),
  gamma.cov = cbind(data$unitcov, data$survcov), epsilon = call(),
  epsilon.cov = cbind(data$unitcov, data$survcov), param = "nu",
  sp.contr = TRUE, modname = NULL, paoname = NULL, outfile,
  model = 3000, fixed = NULL, initvals = NULL, data, VCoutopt = " ",
  noDerived = F, randinit = 0)

Arguments

psi
The right-hand side of the formula for the model to fit for occupancy probability. The terms SP and INT can be used to define a species effect on occupancy and an occurrence-level interaction between species accordingly, without them being defined in psi.cov.

psi.cov
A data frame containing the unit-specific covariates to use for the occupancy component of the model.

p
The right-hand side of the formula for the model to fit for detection probability. The terms SP, INT_o and INT_d can be used to define a species effect on detection, a detection-level interaction where the occurrence of one species changes the detection probability of the other species and a second detection-level interaction where the detection of one species changes the detection probability of the other species in the same survey. These terms do not have to be defined as variables in p.cov.

p.cov
A data frame containing the survey-specific covariates to use for the detection component of the model.

gamma
The right-hand side of the formula for the model to fit for colonization probability. The following terms can be used to define effects on colonization:

SP species effect on colonization (gammaB != gammaA)

SEASON seasonal effect on colonization (gammaXX(t))

INT_Ba interaction of effect on colonization for species B when species A was, or was not present (gammaBA != gammaBa)

INT_B_a interaction of effect on colonization for species B when species A was present in the previous season and was, or was not present in succeeding season (gammaB?A != gammaB?a)

INT_Baa interaction of effect on colonization for species B when species A was not present in the previous season and was, or was not present in succeeding season (gammaBaA != gammaBaa)

INT_Ab interaction of effect on colonization for species A when species B was, or was not present in the previous season (gammaAB != gammaAb)

These terms do not have to be defined as variables in p.cov. For single-season model, set gamma to NULL.

gamma.cov
A data frame containing the site-season-specific covariates to use for the colonization component of the model. This should be a NxK data frame, where N=number of sites, and K=number of sites*number of seasons.

epsilon
The right-hand side of the formula for the model to fit for extinction probability. The terms SP, INT_Ba, INT_B_a, INT_Baa, INT_Ab, SEASON can be used to define effects on extinction, similar to those defined for colonization. These terms do not have to be defined as variables in p.cov.

epsilon.cov
A data frame containing the site-specific covariates to use for the extinction component of the model, similar to gaam.cov.

param
The parameterisation to be used, either "psiBA" or "nu", which relate to type="so.2sp.1" and type="so.2sp.2" in the function occMod respectively.

sp.contr
TRUE (default) or FALSE. Specifies the type of contrast used for the SP term. It it used as the contrasts input in the function contr.treatment.

modname
(optional) a string containing the model name

paoname
(optional) a string containing the filename for the temporary PRESENCE data file.

model
The PRESENCE model code. DO NOT CHANGE.

fixed
A single-column matrix containing values for real parameters to be fixed at. rownnames(fixed) should contain the index of the real parameters to be fixed.

initvals
Initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.

data
The pao data object containing the detection data and other information. NOTE THAT THERE MAY BE SOME CHANGES TO HOW THIS MODEL IS IMPLEMENTED IN THE NEAR FUTURE.

Details

Note: This function assumes data are in "compressed" format (ie.,
0=neither species detected,
1=only species A detected,
2=only species B detected,
3=both species detected)

"Stacked" format can easily be converted to "compressed" format by:
nsites=nrow(det.data)/2
new.det.data=det.data[1:nsites,]+2*det.data[nsites+1:nsites,]

Value

returns a list of class "occMod" and "do.1".

occMod$beta contains the objects:

psi
estimated logistic regression coefficients and standard errors for probability of occurrence in the first year.

psi.VC
variance-covariance matrix for logistic regression coefficients for probability of occurrence in the first year.

p
estimated logistic regression coefficients and standard errors for probability of detection.

p.VC
variance-covariance matrix for logistic regression coefficients for probability of detection.

gamma
estimated logistic regression coefficients and standard errors for probability of colonization.

gamma.VC
var-cov matrix for logistic regression coefficients for probability of colonization.

epsilon
estimated logistic regression coefficients and standard errors for probability of extinction.

epsilon.VC
var-cov matrix for logistic regression coefficients for probability of extinction.

VC
the full variance-covariance matrix for all logistic regression coefficients.

occMod$real contains the objects:

psi
estimated probabilities of occurrence for each sampling unit, along with standard errors and limits of 95% confidence interval. Estimates are provided for the first season (calculated directly from the estimated beta parameters). Estimates for later seasons are provided as derived parameters. The season for which an estimate applies to can be identified from the final number of the row names (rownames(psi)).

p
estimated probabilities of detection for each survey, along with standard errors and limits of 95% confidence interval.

gamma
estimated probabilities of colonization for each sampling unit and season, along with standard errors and limits of 95% confidence interval.

epsilon
estimated probabilities of extinction for each sampling unit and season, along with standard errors and limits of 95% confidence interval.

occMod$derived contains the object:

psi
estimated probabilities of occurrence for each sampling unit for second season onwards, along with standard errors and limits of 95% confidence interval. The season for which an estimate applies to can be identified from the final number of the row names (rownames(psi)).

Author(s)

Darryl MacKenzie

See Also

occMod

Examples

# load a csv data file
filename<-system.file("extdata/twosp_exmpl.csv",package="RPresence")
csv<-read.csv(filename,as.is=T,header=F)
##
nsites=nrow(dethist); nsrvys=ncol(dethist)         #  set number of sites,surveys from det. history data
dethist=matrix(as.integer(unlist(dethist)),nrow=nsites) # replace missing values (-) with NA
cov1=cov2=NULL
##          create input "pao" object, for use with occMod function
data=createPao(dethist,unitcov=cov1,survcov=cov2,title="twosp example")

## fit some models
##    occupancy: species-specific, no interaction, parameters: psiA, psiBA=psiBa
##    detection: species-specific, p=r
mod1<-occMod(model=list(psi~SP,p~SP),data=data,type="so.2sp.1",param="PsiBA")

##    occupancy: species-specific, interaction, parameters: psiA, psiBA, psiBa
##    detection: species-specific, p=r
mod2<-occMod(model=list(psi~SP+INT,p~SP),data=data,type="so.2sp.1",param="PsiBA")

##    occupancy: species-specific, interaction,
##    detection: species=specific, interaction p,r (INT_o) and rBA,rBA (INT_d)
##     Note: this is default 2 sp. model (psiA,psiBA,psiBa,pA,pB,rA,rBA,rBa)
mod3<-occMod(model=list(psi~SP+INT,p~SP+INT_o+INT_d+SP:INT_o),data=data,type="so.2sp.1",param="PsiBA")
#
## Not run: 
m1<-occMod(model=list(psi~SP,p~SP,data=data,type="so.2sp.1",param="PsiBA")

## fit some multi-season, 2-species models (using different data)
mod0<-occMod(model=list(psi=psi~SP+INT,                                   #  PsiA,psiB,psiBa
                         gamma =gamma~SP+INT_Ab+INT_Ba+INT_B_a+INT_Baa,    #  gamAB,gamAb,gamBAA,gamBAa,gamBaA,gamBaa
                         epsilon=epsilon~SP+INT_Ab+INT_Ba+INT_B_a+INT_Baa, #  epsAB,epsAb,epsBAA,epsBAa,epsBaA,epsBaa
                         p      =p~SP+INT_o+INT_d+SP:INT_o),               #  pA,pB,rA,rBA,rBa
              data=data,type="do.2sp.1")
mod1<-occMod(model=list(psi=psi~SP+INT,                                   #  PsiA,psiB,psiBa
                         gamma =gamma~SP+INT_Ab+INT_Ba,                    #  gamAB,gamAb,gamBA.,gamBa.
                         epsilon=epsilon~SP+INT_Ab+INT_Ba,                 #  epsAB,epsAb,epsBA.,epsBa.
                         p      =p~SP+INT_o+INT_d+SP:INT_o),               #  pA,pB,rA,rBA,rBa
              data=data,type="do.2sp.1")
mod2<-occMod(model=list(psi=psi~SP+INT,                                   #  PsiA,psiB,psiBa
                         gamma =gamma~SP,                                  #  gamA.,gamB..
                         epsilon=epsilon~SP,                               #  epsA.,epsB..
                         p      =p~SP+INT_o+INT_d+SP:INT_o),               #  pA,pB,rA,rBA,rBa
              data=data,type="do.2sp.1")

tbl=create_aic_table(list(mod0,mod1,mod2))
print(tbl$table)

## End(Not run)

[Package RPresence version 2.12.17 Index]