occMod_2SPfp {RPresence}	R Documentation

Fit two species, static occupancy (single season) model with false positive detections

Description

This is not intended for direct use, but instead the occMod function should be used with type="so.2sp.cd". NOTE THAT THERE MAY BE SOME CHANGES TO HOW THIS MODEL IS IMPLEMENTED IN THE NEAR FUTURE.

Usage

occMod_2SPfp(psi = call(), psi.cov = data$unitcov, p = call(),
  p.cov = cbind(data$unitcov, data$survcov), omeg = call(), omeg.cov,
  conf = call, conf.cov, param = "PsiBA", sp.contr = TRUE,
  modname = NULL, paoname = NULL, outfile, model = 3000,
  maxfn = "32000", fixed = NULL, initvals = NULL, data, randinit = 0,
  VCoutopt = " ", noDerived = F)

Arguments

psi
The right-hand side of the formula for the model to fit for occupancy probability. The terms SP and INT can be used to define a species effect on occupancy and an occurrence-level interaction between species accordingly, without them being defined in psi.cov.

psi.cov
A data frame containing the unit-specific covariates to use for the occupancy component of the model.

p
The right-hand side of the formula for the model to fit for detection probability. The terms SP, INT_o and INT_d can be used to define a species effect on detection, a detection-level interaction where the occurrence of one species changes the detection probability of the other species and a second detection-level interaction where the detection of one species changes the detection probability of the other species in the same survey. These terms do not have to be defined as variables in p.cov.

p.cov
A data frame containing the survey-specific covariates to use for the detection component of the model.

param
The parameterisation to be used, either "psiBA" or "nu", which relate to type="so.2sp.1" and type="so.2sp.2" in the function occMod respectively.

sp.contr
TRUE (default) or FALSE. Specifies the type of contrast used for the SP term. It it used as the contrasts input in the function contr.treatment.

modname
(optional) a string containing the model name

paoname
(optional) a string containing the filename for the temporary PRESENCE data file.

model
The PRESENCE model code. DO NOT CHANGE.

fixed
A single-column matrix containing values for real parameters to be fixed at. rownnames(fixed) should contain the index of the real parameters to be fixed.

initvals
Initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.

data
The pao data object containing the detection data and other information. NOTE THAT THERE MAY BE SOME CHANGES TO HOW THIS MODEL IS IMPLEMENTED IN THE NEAR FUTURE.

theta
The right-hand side of the formula for the model to fit for "local-use" probability. The following terms are pre-defined covariates to allow constraints on the theta's (thA, thA\', thBA, thBA\', thBa, thBa\').

SP - species effect on theta's (thA != thB),

PRIME - "PRIME" effect (thA !=thA\', thB. !=thB.\')

BAa - species interaction on theta (thBA != thBa, thBA\' != thBa\')

INTth - interaction of PRIME and BAa (thBA != thBA\' and thBA != thBa)

So, the most generl model (all theta\'s different) would be: theta(~SP*PRIME+BAa+INTth)

theta.cov
A data frame containing the survey-specific covariates to use for the theta component of the model.

Value

returns a list of class occMod and so2spCd

occMod$beta contains the objects:

psi
estimated logistic regression coefficients and std.err for prob. of occurrence.

psi.VC
var-covar matrix for logistic regression coefficients for prob. of occurrence.

p
estimated logistic regression coefficients and standard errors for probability of detection.

p.VC
variance-covariance matrix for logistic regression coefficients for probability of detection.

omeg
estimated logistic regression coefficients and std.err for prob. of local occurrence before 1st survey.

omeg.VC
var-covar matrix for omeg.

conf
estimated logistic regression coefficients and std.err for prob. of sample is confirmed.

conf.VC
var-covar matrix for conf.

VC
the full variance-covariance matrix for all logistic regression coefficients.

occMod$real contains the objects:

psi
estimated probabilities of occurrence for each sampling unit, along with standard errors and limits of 95% confidence interval.

theta
estimated prob. of local occurrence in each survey w/ std.err and 95% conf interval limits.

p
estimated probabilities of detection for each survey, along with standard errors and limits of 95% confidence interval.

th0pi
estimated prob. of local occurrence before 1st survey w/ std err and 95% conf. interval limits.

Author(s)

Jim Hines and Darryl MacKenzie

See Also

occMod

Examples

## Not run: 
# load a csv data file
filename<-system.file("extdata/twosp_exmpl.csv",package="RPresence")
dethist<-read.csv(filename,as.is=T)

nsites=nrow(dethist); nsrvys=ncol(dethist)         #  set number of sites,surveys from det. history data
dethist=matrix(as.integer(unlist(dethist)),nrow=nsites) # replace missing values (-) with NA

##          create input "pao" object, for use with occMod function
data=createPao(dethist,unitcov=NULL,survcov=NULL,title="twosp corr.det. example")

## fit some models

mod1<-occMod(model=list(psi~SP,   # occupancy species-specific, no interaction, parameters: psiA, psiBA=psiBa
                        theta~SP, # local occ. species-specific: parameters, thetaA0, thetaA1 (no corr. det. model)
                        p~SP,     # detection sp. specific: parms: pA, pB
                        th0pi~1),data=data,type="so.2sp.cd",param="PsiBA")

mod2=occMod(model=list(psi~SP+INT,    # species and interaction effect (psiA != psiBA != psiBa)
                       theta~SP*PRIME+BAa+INTth,   # thA != thA' != thBA != thBA' != thBa != thBa'
                       p~SP+INT_o+INT_d+INT_so,    # pA != pB != rA != rBA !=rBa
                       th0pi~1),                   # constant prop unocc before 1st segment
            data=data, type="so.2sp.cd", fixed=NULL)
#
tbl=createAicTable(list(mod1,mod2)); print(tbl$table)

sim_2sp_corr_det_data <- function(sites=100,surveys=12,psiA=.8,psiBA=.3,psiBa=.7,
                                  thA=c(.4,.9),thBA=c(.3,.9),thBa=c(.3,.9),  #  1st is theta, 2nd is theta\'
                                  pA=.6,pB=.66,rA=.7,rBA=.4,rBa=.5,
                                  th0piA=0,th0piBA=0,th0piBa=0) {
              #  simulates single-season 2-species correlated detections data
 psiB=c(psiBa,psiBA); th0piB=c(th0piBa,th0piBA); h=matrix(0,sites,surveys)
 occA=0+(runif(sites)<psiA);               # randomly assign occupancy state for sp. A of each site (0=unocc, 1=occ)
 occB=0+(runif(sites)<psiB[occA+1])        # randomly assign occupancy state for sp. B; depends on occupancy of sp. A
 loccA=occA*(runif(sites)<th0piA);         # randomly assign local-occ state before 1st segment of trail for sp. A
 loccB=occB*(runif(sites)<th0piB[loccA+1]) # randomly assign local-occ state before 1st segment of trail for sp. B
 thB=c(thBa,thBA); pAx=c(pA,rA); pBx=c(pB,rBa,NA,rBA)
 for (i in 1:surveys) {
   loccA=occA*(runif(sites)<thA[loccA+1]);       # local-occ for survey i depends on previous local-occ
   loccB=occB*(runif(sites)<thB[loccB+1+2*occA])
   detA=loccA*(runif(sites)<pAx[occB+1])         # detection depends on local-occ state (p=0 if not local-occ)
   detB=loccB*(runif(sites)<pBx[occA+1+2*detA])
   h[,i]=detA+2*detB  #  detection-history in compressed format: 0=neither sp. detected, 1=sp. A only, 2=sp. B only, 3=both
 }
 return(h)
}

x=sim_2sp_corr_det_data(1000)          # simulate some data
data=createPao(x)                     # create_pao file from data
#    run a model : psiA()psiBA(),psiBa,thA(),thA'(),thBA(),thBA'(),thBa(),thBa'(),pA(),pB(),rA(),rBA(),rBa()
mod3=occMod(model=list(psi~SP+INT,    # species and interaction effect (psiA != psiBA != psiBa)
          theta~SP*PRIME+BAa+INTth,   # thA != thA' != thBA != thBA' != thBa != thBa'
          p~SP+INT_o+INT_d+INT_so,    # pA != pB != rA != rBA !=rBa
          th0pi~1),                   # constant prop unocc before 1st segment
           data=data, type="so.2sp.cd", fixed=NULL)
print_one_site_estimates(mod3,site=1)

## End(Not run)

[Package RPresence version 2.12.17 Index]