occMod_DO4 {RPresence}	R Documentation

Fit dynamic (multi-season) occupancy model using the fourth parameterisation (random changes in occupancy, epsilon=1-gamma) in PRESENCE.

Description

This is not intended for direct use, but instead the occMod function should be used with type="do.4".

Usage

occMod_DO4(psi = call(), psi.cov = data$unitcov, p = call(),
  p.cov = cbind(data$unitcov, data$survcov), modname = NULL,
  paoname = NULL, outfile, model = 240, fixed = NULL, initvals = NULL,
  data, conf, VCoutopt = " ", noDerived = F, randinit = 0)

Arguments

psi
the right-hand side of the formula for the model to fit for occupancy probability in each season.

psi.cov
a data frame containing the unit-specific covariates to use for the occupancy component of the model, with number of rows = data$nunits or data$nunits*data$nseasons. If the shorter version of the data frame is supplied the rows are recycled to the longer length.

p
the right-hand side of the formula for the model to fit for detection probability.

p.cov
a data frame containing the survey-specific covariates to use for the detection component of the model.

modname
(optional) a string containing the model name

paoname
(optional) a string containing the filename for the temporary PRESENCE data file.

fixed
a single-column matrix containing values for real parameters to be fixed at. rownnames(fixed) should contain the index of the real parameters to be fixed.

initvals
initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.

data
the pao data object containing the detection data and other information.

conf
limits for confidence intervals as a proportion (defalut=0.95 for 95% conf. interval limits)

Value

list of class "occMod" and "do4".

occMod$beta contains the objects:

psi
estimated logistic regression coefficients and standard errors for probabilities of occurrence in each season.

psi.VC
variance-covariance matrix for psi.

p
estimated logistic regression coefficients and standard errors for probabilities of detection in each survey.

p.VC
variance-covariance matrix for p.

VC
the full variance-covariance matrix for all logistic regression coefficients.

occMod$real contains the objects:

psi
estimated probabilities of occurrence for each sampling unit, in each season, along with standard errors and limits of 95% confidence interval. The season for which an estimate applies to can be identified from the final number of the row names (rownames(psi)).

p
estimated probabilities of detection for each survey, along with standard errors and limits of 95% confidence intervals.

occMod$derived contains the objects:

gamma
estimated probabilities of colonization for each sampling unit, along with standard errors and limits of 95% confidence intervals. The beginning season for which an estimate applies to can be identified from the final number of the row names (rownames(gamma)).

epsilon
estimated probabilities of extinction for each sampling unit, along with standard errors and limits of 95% confidence intervals. The beginning season for which an estimate applies to can be identified from the final number of the row names (rownames(epsilon)). Under this parameterisation, extinction probabilities have the same value as 1-colonization probabilties

Note: For this parameterization, gamma(t)=psi(t+1), due to the constraint, epsilon=1-gamma

           ie., psi(t+1)=psi(t)*(1-epsilon(t))+(1-psi(t))*gamma(t)
                psi(t+1)=psi(t)*gamma(t)+(1-psi(t))*gamma(t)
                psi(t+1)=psi(t)*gamma(t)+gamma(t)-psi(t)*gamma(t)
                psi(t+1)=gamma(t)

Author(s)

Darryl MacKenzie

See Also

occMod

[Package RPresence version 2.12.17 Index]