occMod_DO_fp {RPresence}	R Documentation

Fit static occupancy/single season model

Description

This is not intended for direct use, but instead the occMod function should be used with type="do.fp".

Usage

occMod_DO_fp(psi = call(), psi.cov = data$unitcov, gamma = call(),
  gamma.cov = data$unitcov, epsilon = call(), epsilon.cov = data$unitcov,
  p11 = call(), p11.cov = cbind(data$unitcov, data$survcov), p10 = call(),
  p10.cov = cbind(data$unitcov, data$survcov), b = call(),
  b.cov = cbind(data$unitcov, data$survcov), modname = NULL,
  paoname = NULL, outfile, model = 100, fixed = NULL, initvals = NULL,
  data, conf, VCoutopt = " ", noDerived = F, randinit = 0)

Arguments

psi
the right-hand side of the formula for the model to fit for occupancy probability.

psi.cov
a data frame containing the unit-specific covariates to use for the occupancy component of the model.

gamma
the right-hand side of the formula for the model to fit for colonization probability in each season.

gamma.cov
a data frame containing the unit-specific covariates to use for the colonization component of the model, with number of rows = data$nunits*(data$nseasons-1).

epsilon
the right-hand side of the formula for the model to fit for extinction probability in each season.

epsilon.cov
a data frame containing the unit-specific covariates to use for the extinction component of the model, with number of rows = data$nunits*(data$nseasons-1).

p11
the right-hand side of the formula for the model to fit for "sure" detection probability.

p11.cov
a data frame containing the survey-specific covariates to use for the detection component of the model.

p10
the right-hand side of the formula for the model to fit for false-positive detection probability.

p10.cov
a data frame containing the survey-specific covariates to use for the detection component of the model.

b
the right-hand side of the formula for the model to fit for Pr(detection is "sure")

b.cov
a data frame containing the survey-specific covariates to use for the b

modname
(optional) a string containing the model name

paoname
(optional) a string containing the filename for the temporary PRESENCE data file.

model
the PRESENCE model code. DO NOT CHANGE.

fixed
a single-column matrix containing values for real parameters to be fixed at. rownnames(fixed) should contain the index of the real parameters to be fixed.

initvals
initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.

data
the pao data object containing the detection data and other information.

conf
level for confidence interval (may be vector valued).

Value

returns a list of class "occMod" and "soFp".

occMod$beta contains the objects:

psi
estimated logistic regression coefficients and standard errors for probability of occurrence.

psi.VC
variance-covariance matrix for logistic regression coefficients for probability of occurrence.

gamma
estimated logistic regression coefficients and standard errors for probability of colonization.

gamma.VC
variance-covariance matrix for logistic regression coefficients for probability of colonization.

epsilon
estimated logistic regression coefficients and standard errors for probability of extinction.

epsilon.VC
variance-covariance matrix for logistic regression coefficients for probability of extinction.

p
estimated logistic regression coefficients and standard errors for probability of detection.

p.VC
variance-covariance matrix for logistic regression coefficients for probability of detection.

VC
the full variance-covariance matrix for all logistic regression coefficients.

occMod$real contains the objects:

psi
estimated probabilities of occurrence for each sampling unit, along with standard errors and limits of 95% confidence interval.

p
estimated probabilities of detection for each survey, along with standard errors and limits of 95% confidence interval.

occMod$derived contains the objects:

psi_c
estimated probabilities of occurrence given the detection history for each sampling unit, along with standard errors and limits of 95% confidence interval. Will be =1 for any unit where the species was detected at least once.

Author(s)

Darryl MacKenzie

See Also

occMod, pao

[Package RPresence version 2.12.17 Index]