occMod_DO_ms1 {RPresence}	R Documentation

Fit multi-state dynamic (multi-season) occupancy model using the 1st parameterisation (ψ/φ).

Description

This is not intended for direct use, but instead the occMod function should be used with type="do.ms.1".

Usage

occMod_DO_ms1(psi = call(), psi.cov = data$unitcov, phi = call(),
  phi.cov = data$unitcov, p = call(), p.cov = data$survcov,
  modname = NULL, paoname = NULL, outfile, model = 6000, fixed = NULL,
  initvals = NULL, data, VCoutopt = " ", noDerived = F, randinit = 0)

Arguments

psi
the right-hand side of the formula for the model to fit for occupancy probability.

psi.cov
a data frame containing the unit-specific covariates to use for the occupancy component of the model.

phi
the right-hand side of the formula for the model to fit for occupancy transition probabilities,.

phi.cov
a data frame containing the unit-specific covariates to use for the occupancy component of the model.

p
the right-hand side of the formula for the model to fit for detection probability.

p.cov
a data frame containing the survey-specific covariates to use for the detection component of the model.

modname
(optional) a string containing the model name

paoname
(optional) a string containing the filename for the temporary PRESENCE data file.

model
the PRESENCE model code. DO NOT CHANGE.

fixed
a 2-column matrix containing indices and values for real parameters to be fixed at.

initvals
initial values for the beta parameters at which PRESENCE begins the optimisation. Default = 0.

data
the pao data object containing the detection data and other information.

Value

returns a list of class "occMod" and "do.1".

occMod$beta contains the objects:

psi
estimated logistic regression coefficients and standard errors for probability of occurrence in the first year.

psi.VC
variance-covariance matrix for logistic regression coefficients for probability of occurrence in the first year.

phi
estimated logistic regression coefficients and standard errors for probability of state transition between years.

phi.VC
variance-covariance matrix for logistic regression coefficients for probability of transition between years.

p
estimated logistic regression coefficients and standard errors for probability of detection.

p.VC
variance-covariance matrix for logistic regression coefficients for probability of detection.

VC
the full variance-covariance matrix for all logistic regression coefficients.

occMod$real contains the objects:

psi
estimated probabilities of occurrence for each sampling unit and state, along with standard errors and limits of 95% confidence interval.

phi
estimated probabilities of transition between occupancy states for each sampling unit and state, along with standard errors and limits of 95% confidence interval.

p
estimated probabilities of detection for each survey, along with standard errors and limits of 95% confidence interval.

Note

NOTE THAT THIS FUNCTION HAS NOT BEEN EXTENSIVELY TESTED SO EXACT IMPLEMENTATION MAY CHANGE.

psi built-in covariates:

STATE - for model where psi (initial occupancy) depends on occupancy state.

phi built-in covariates:

FROM - for model where psi depends on state species is transitioning from.

TO - for model where psi depends on state species is transitioning to.

FROM_TO - for model where psi depends on "from" state and "to" state (equalivent to "FROM*TO").

SEASON - for model where psi changes from season to season.

P built-in covariates:

OSTATE - detection depends on observed state of species.

OTSTATE - detection depends on both observed and true state of species.

SURVEY - detection varies by survey.

SEASON - detection varies by season (constant within season).

Note: P21 (prob detect as state=2, when true state=1) must be fixed to 0, or model will not converge Also: setting outfile="modname" causes RPresence to save the output file using the model name (with ".out" extension). Otherwise, you can name it something else. Or, set outfile=NULL to not save the output file.

Author(s)

Jim Hines

See Also

occMod, occMod_DO_ms2

[Package RPresence version 2.12.17 Index]