occMod_DO_ms2 {RPresence}	R Documentation

Fit multi-state dynamic occupancy (multi-season) occupancy model using the second (conditional binomial) parameterisation in PRESENCE.

Description

This is not intended for direct use, but instead the occMod function should be used with type="do.ms.2".

Usage

occMod_DO_ms2(psi = call(), psi.cov = data$unitcov, r = call(),
  r.cov = data$unitcov, p = call(), p.cov = data$survcov,
  delta = call(), delta.cov = data$survcov, modname = NULL,
  paoname = NULL, outfile, model = 6100, fixed = NULL, initvals = NULL,
  data, VCoutopt = " ", noDerived = F, randinit = 0)

Arguments

psi
the right-hand side of the formula for the model to fit for occupancy probability, both first year and dynamic occupancy. The terms SEASON, DYN and PREV_STATE can each be used without having been defined in psi.cov. SEASON allows for a seasonal effect, DYN enables dynamic occupancy probabilities to be different from first year occupancy, and PREV_STATE allows the probability to be different depending on the state in the previous season.

psi.cov
a data frame containing the unit-specific covariates to use for the occupancy component of the model.

r
the right-hand side of the formula for the model to fit for the probability of being in the second state, conditional on the unit being occupied. The terms SEASON, DYN and PREV_STATE can be used as above with the same effect.

r.cov
a data frame containing the unit-specific covariates to use for the probability of being in the second state, conditional on the unit being occupied part of the model.

p
the right-hand side of the formula for the model to fit for detection probability. The terms SEASON and STATE can each be used without having been defined in p.cov. SEASON allows for a seasonal effect, and STATE allows the probability of detection to be different depending on the state in the current season.

p.cov
a data frame containing the survey-specific covariates to use for the detection component of the model.

delta
the right-hand side of the formula for the model to fit for the probability of detecting the second state in a survey, conditional on the species being detected in the survey. The terms SEASON and STATE can each be used as above to the same effect.

delta.cov
a data frame containing the survey-specific covariates to use for the probability of detecting the second state in a survey, conditional on the species being detected in the survey.

modname
(optional) a string containing the model name

paoname
(optional) a string containing the filename for the temporary PRESENCE data file.

model
the PRESENCE model code. DO NOT CHANGE.

fixed
a single-column matrix containing values for real parameters to be fixed at. rownnames(fixed) should contain the index of the real parameters to be fixed.

initvals
initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.

data
the pao data object containing the detection data and other information.

NOTE THAT THIS FUNCTION HAS NOT BEEN EXTENSIVELY TESTED SO EXACT IMPLEMENTATION MAY CHANGE.

Value

returns a list of class "occMod" and "do.1".

occMod$beta contains the objects:

psi
estimated logistic regression coefficients and standard errors for probability of occurrence in the first year.

psi.VC
variance-covariance matrix for logistic regression coefficients for probability of occurrence in the first year.

p
estimated logistic regression coefficients and standard errors for probability of detection.

p.VC
variance-covariance matrix for logistic regression coefficients for probability of detection.

VC
the full variance-covariance matrix for all logistic regression coefficients.

occMod$real contains the objects:

psi
estimated probabilities of occurrence for each sampling unit, along with standard errors and limits of 95% confidence interval. Estimates are provided for the first season (calculated directly from the estimated beta parameters). Estimates for later seasons are provided as derived parameters. The season for which an estimate applies to can be identified from the final number of the row names (rownames(psi)).

p
estimated probabilities of detection for each survey, along with standard errors and limits of 95% confidence interval.

occMod$derived contains the objects:

psi
estimated probabilities of occurrence for each sampling unit for second season onwards, along with standard errors and limits of 95% confidence interval. The season for which an estimate applies to can be identified from the final number of the row names (rownames(psi)).

Author(s)

Darryl MacKenzie

See Also

occMod

[Package RPresence version 2.12.17 Index]