occMod_SO_cd {RPresence}	R Documentation

Fit static occupancy, correlated detections model

Description

This is not intended for direct use, but instead the occMod function should be used with type="so.cd".

Usage

occMod_SO_cd(psi = call(), psi.cov = data$unitcov, theta = call(),
  theta.cov = data$unitcov, p = call(), p.cov = data$survcov,
  th0pi = call(), th0pi.cov = data$unitcov, modname = NULL,
  paoname = NULL, outfile, model = 103, fixed = NULL, initvals = NULL,
  data, VCoutopt = " ", noDerived = F, randinit = 0)

Arguments

psi
the right-hand side of the formula for the model to fit for occupancy probability (eg., ~1).

psi.cov
a data frame containing the unit-specific covariates to use for the occupancy component of the model.

theta
the right-hand side of the formula for the model to fit for secondary-scale occupancy probability. (eg., ~PRIME -> th0 != th1, ~SURVEY -> th0(1)=th1(1), th0(2)=th1(2),...)

theta.cov
a data frame containing the secondary-scale specific covariates to use for the secondary-scaleoccupancy component of the model.

p
the right-hand side of the formula for the model to fit for detection probability.

p.cov
a data frame containing the survey-specific covariates to use for the detection component of the model.

th0pi
the right-hand side of the formula for the model to fit for local use before 1st segment.

th0pi.cov
a data frame containing the unit-specific covariates to use for th0pi.

modname
(optional) a string containing the model name

paoname
(optional) a string containing the filename for the temporary PRESENCE data file.

model
the PRESENCE model code. DO NOT CHANGE.

fixed
a single-column matrix containing values for real parameters to be fixed at. rownnames(fixed) should contain the index of the real parameters to be fixed.

initvals
initial values for the beta parameters at which PRESENCE begins the optimisation. The default values in PRESENCE is 0.

data
the pao data object containing the detection data and other information.

Value

list of class "occMod" and "soCd".

occMod$beta contains the objects:

psi
estimated logistic regression coefficients and standard errors for probability of occurrence.

psi.VC
variance-covariance matrix for logistic regression coefficients for probability of occurrence.

theta
estimated logistic regression coefficients and standard errors for probability of occurrence at the secondary scale.

theta.VC
variance-covariance matrix for logistic regression coefficients for probability of occurrence at the secondary scale.

p
estimated logistic regression coefficients and standard errors for probability of detection.

p.VC
variance-covariance matrix for logistic regression coefficients for probability of detection.

th0pi
estimated logistic regression coefficients and standard errors for probability of local-use before 1st segment.

th0pi.VC
variance-covariance matrix for logistic regression coefficients for th0pi.

VC
the full variance-covariance matrix for all logistic regression coefficients.

occMod$real contains the objects:

psi
estimated probabilities of occurrence for each sampling unit, along with standard errors and limits of 95% confidence interval.

theta
estimated probabilities of occurrence at the secondary scale for each sampling unit, along with standard errors and limits of 95% confidence interval.

p
estimated probabilities of detection for each survey, along with standard errors and limits of 95% confidence interval.

psi_c
estimated probabilities of occurrence given the detection history for each sampling unit, along with standard errors and limits of 95% confidence interval. Will be =1 for any unit where the species was detected at least once.

Examples

#   example 1 - read previously created input file and run a model
data=readPao('genpres.pao')  #  read simulated data from program GENPRES
mod1=occMod(model=list(psi~1, theta~PRIME, p~1, th0pi~1), data=data, type="so.cd")

#   example 2 - simulate data and run a mode
sim_corr_det_data <- function(sites=500,surveys=9,psi=.9,th=c(.2,.8),p=.6,th0pi=0) {
  # simulates single-season correlated detections data
  h=matrix(0,sites,surveys); px=c(0,p) # px=detection prob for each state (local-unocc, local-occ)
  occ=0+(runif(sites)<psi);            # randomly assign occupancy state of each site (0=unocc, 1=occ)
  locc=occ*(runif(sites)<th0pi);       # randomly assign local-occ state before 1st segment of trail
  for (i in 1:surveys) {
     locc=occ*(runif(sites)<th[locc+1]);   # local-occ for survey i depends on previous local-occ
     h[,i]=locc*(runif(sites)<px[locc+1])  # detection depends on local-occ state (p=0 if not local-occ)
  }
return(h)
}
x=sim_corr_det_data(sites=1000,surveys=15)
data=createPao(x)
mod1=occMod(model=list(psi~1,       # constant psi
                     theta~PRIME, # theta not equal theta'
                     p~1,         # constant detection
                     th0pi~1      # constant prop unocc before 1st segment
), data=data, type="so.cd")  #  run ss-cd model
print_one_site_estimates(mod1,site=1)

[Package RPresence version 2.12.17 Index]