predict.occMod {RPresence}	R Documentation

Predict real parameter values.

Description

Predict real parameter values from a model that has been fitted to the data, using the variable or covariates values that are supplied by the user. Currently has only been coded to predict probabilities and checked for the static, or single-season, occupancy model.

Usage

## S3 method for class 'occMod'
predict(object, newdata, param = NULL, conf = 0.95)

Arguments

object
and object of class occMod which is the model that has been fitted to the data from which predictions are required. The result of a call to occMod.

newdata
a data frame containing the values of the variables for which predictions are required. Variable names must exactly match those used in the fitted model. May contain additional variables not used in the fitted model.

param
a string containing the name of the real parameter for which predictions are required, e.g., type="psi".

conf
level for confidence interval (may be vector valued).

Details

Note that can not make predictions directly from models that have been imported from a PRESENCE project using the importProject function.

Value

returns a data frame containing the predicted value, standard error and limits of a 95% confidence interval.

Author(s)

Darryl MacKenzie

See Also

occMod

Examples

psi.real=predict(mod1,salmdr.data$survcov,param="psi",conf=0.95); print(head(psi.real))

[Package RPresence version 2.12.17 Index]