print_one_site_estimates {RPresence}	R Documentation

Print real param estimates for i-th site only print_one_site_estimates Prints real param estimates for ith site only

Description

Print real param estimates for i-th site only print_one_site_estimates Prints real param estimates for ith site only

Usage

print_one_site_estimates(mod, site = 1)

Arguments

mod
model output from occMod function.

site
site number(s) to print (default=1)

Value

nothing

[Package RPresence version 2.12.17 Index]