readPao {RPresence}	R Documentation

Read a pao data object.

Description

readPao Creates a pao data object from a PRESENCE data file.

Usage

readPao(f)

Arguments

f
A string specifying the name of the PRESENCE data file.

Value

list of class "pao" containing the data required to fit an occupancy model using occMod. The list contains the following objects:

nunits
number of units or sampling units, i.e., nrow(data).

nsurveys
total number of surveys, i.e., ncol(data).

nseasons
number of seasons of data.

det.data
detection/nondetection data, i.e., data.

unitcov
see above. If no covariates are supplied a vector of 1's named TEMP is added.

survcov
see above. A factor called SURVEY is automatically added to allow survey-specific detection probabilities (but equal across units).

nsurveyseason
see above.

title
see above.

unitnames
see above.

surveynames
a vector of strings providing labels for the survey occasions. Of little practical consequence.

paoname
see above. Equals f for createPao.

Author(s)

Darryl MacKenzie and Jim Hines

See Also

occMod

Examples

readPao(system.file('extdata/skinks.pao',package='RPresence'))

[Package RPresence version 2.12.17 Index]