runPresence {RPresence}	R Documentation

call PRESENCE to produce model estimates (replaces "write_dm_and_run")

Description

runPresence calls PRESENCE to produce model estimates.

Usage

runPresence(a, dm = list(), model, modname = "model1", fixedvals = NULL,
  initvals = NULL, modfitboot = NULL, maxfn = 32000, limit.real = FALSE,
  VCoutopt = " ", noDerived = FALSE, randinit = 0, outfile)

Arguments

a
pao object

model
model number (not used anymore)

modname
name of model (eg., "psi(.)p(.)")

initvals
vector of beta initial values

maxfn
maximum number if iterations for optimization routine

limit.real
causes PRESENCE to only print estimates for the 1st site

VCoutopt
option for PRESENCE computation/output of beta/real var-cov matrices

noDerived
if TRUE, doesn't print derived estimates from model

dms
list of design matrices (should be 6 in list)

fixed
data.frame of fixed values (e.g., fixed$params<-c("p[1]","p[3]"),fixed$value<-0)

Details

Input is passed to PRESENCE via arguments (no pao or design matrix input files needed). There are three choices for output:

outfile=NULL - no output file saved, results in R variables only,

outfile="modname" - output file saved with name of model and ".out" appended,

outfile="xyz" - output file saved with user-specified name, (eg., "xyz.out")

The first option is best for running simulations of a model using "Snowfall" for multi-tasking. The 2nd option is useful when you would like to find the full text output of a model from Windows explorer. The 3rd option allows you to specify any output file name.

choices for VCoutopt are:

"nose"only compute likelihood and beta estimates, no beta var-cov matrix or beta SE's (useful for faster model selection of big models or simulations)

"betavc"compute but don't print beta var-cov matrix so we get beta SE's, don't compute real params

"noreal"compute and print beta beta var-cov matrix, don't compute real params (I think this is what you want for RPresence)

"(default)"compute real params and real var-cov matrix, don't print either var-cov matrix.

"realbetavc"compute real params and real var-cov matrix, print only beta var-cov matrix.

"bothvc"print both var-cov matrices.

Value

list(LL=log-likelihood, beta=beta estimates, betaVC=var-cov matrix of betas).

Author(s)

Darryl MacKenzie and Jim Hines

See Also

occMod

Examples

runPresence("test.pao",list(dm1,dm2,dm3,dm4,dm5,dm6),100,"psi(.)p(.)")

[Package RPresence version 2.12.17 Index]