simplify_dm {RPresence}	R Documentation

simplify design matrix by replacing all individual covariates which are all one or zero with '1' or '0' in the design matrix. This makes the design matrix much easier to read and faster for the model. simplify_dm Replaces ind. covariates which are all 0 or 1, with 0 or 1.

Description

simplify design matrix by replacing all individual covariates which are all one or zero with '1' or '0' in the design matrix. This makes the design matrix much easier to read and faster for the model. simplify_dm Replaces ind. covariates which are all 0 or 1, with 0 or 1.

Usage

simplify_dm(dm, pao)

Arguments

dm
model design matrix.

pao
pao file created by occMod.???

Value

simplified design matrix

[Package RPresence version 2.12.17 Index]