summary.aicTable {RPresence}	R Documentation

Summary of aic.table object

Description

Provides a brief summary of an object of class aicTable, which is the result of the function createAicTable.

Usage

## S3 method for class 'aicTable'
summary(object, ...)

Arguments

object
an object of class aicTable which is the result of the function createAicTable

...
possible additional arguments

Value

Produces a summary of the model comparison using AIC as a data frame that includes the model name, relative difference in AIC, model weight, number of parameters, twice the negative log-likelihood and any warnings.

Author(s)

Darryl MacKenzie

[Package RPresence version 2.12.17 Index]