summary.occMod {RPresence}	R Documentation

Summary of occupancy model

Description

Provides a brief summary of an object of class occMod, which is the result from fitting amodel using the occMod function.

Usage

## S3 method for class 'occMod'
summary(object, ...)

Arguments

object
an object of class occMod which is the result from fitting a model using the occMod function.

...
possible additional arguments

Value

Provides a brief summary of the fitted model, including model name, AIC value and number of parameters.

Author(s)

Darryl MacKenzie

[Package RPresence version 2.12.17 Index]