summary.pao {RPresence}	R Documentation

Summary of occupancy data object

Description

Provides a brief summary of an object of class pao, which is the input data for occupancy analysis.

Usage

## S3 method for class 'pao'
summary(object)

Arguments

object
an object of class pao created by readPao or createPao.

Value

Provides a brief summary of the input object, including number of sites, surveys.

Author(s)

Jim Hines

[Package RPresence version 2.12.17 Index]