summedWgt {RPresence}	R Documentation

Calculate summed model weights for a set of variables for a specific parameter.

Description

Calculates the summed model weights (and evidence ratios) for a set of variables covnames, for a parameter param and a set of models stored in an aic.tab object.

Usage

summedWgt(covnames, param, aic.tab)

Arguments

covnames
a vector of string values containing the names of the variables of interest.

param
a string with the name of the parameter of interest, e.g., param="psi".

aic.tab
an aic.tab object that includes the set of models from which the summed model weights are to be calculated.

Value

returns a data frame with the summed weights and evidence ratios.

Author(s)

Darryl MacKenzie.

See Also

createAicTable

[Package RPresence version 2.12.17 Index]